dns-proxy: Don't restart IPv6 rules when IPv6 addr is not changed
This patch stops DNS proxy from stopping and starting IPv6 rules
when RTNL gives the same IPv6 address. One reason this can happen
is if IPv6 inside the namespace is disabled and re-enabled. This
might be a reason of a log spam reported in b/197267183.

BUG=b:197267183
TEST=Multiple same RTNL callbacks does not restart IPv6 rules.

Change-Id: Ia822e08e1b6bbd9b1457a2c053f4e8ee5a021692
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3277616
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>