spaced: Add seccomp policies for pvdisplay, VM builders
Add seccomp policies for pvdisplay to account for using the lock
directory and for getting the physical volume for the stateful
partition.

Additionally, VM tests (direct-tast-vm) report failures on
gettimeofday and clock_gettime that are not seen on regular
devices. Add these as well.

BUG=chromium:1268962
TEST=boot and query from spaced, lvm no longer crashes.

Change-Id: I6ae5e942b46e9c4341d8c55c68628964e9d31daf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3276658
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>