camera: add a tool for modifying on-device setting files
The config_file_ui tool can be used to host a webserver and expose UIs
for various setting options. The UI and options can be customized using
JSON config files. The setting options are automatically updated and
written to the designated config files on the device when modified on
the UI.

Also add the config descriptor files for the existing feature tuning
options.

BUG=None
TEST=Manually on device

Change-Id: I7629d6f9c8d77ecc1d520c4934820b51b61281ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3288517
Reviewed-by: Shik Chen <shik@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>