patchpanel: fix broadcast_forwarder buffer overflow
This patch fixes a buffer overflow issue in BroadcastForwarder's handler
for RTNL RTM_NEWADDR messages. Even though these binary messages come
from the kernel and could be trusted in theory, it is better to not
assume that the netlink attribute sizes directly parsed from the
messages exactly fit the in-memory fields where the message data is
copied.

BUG=b:203712717
BUG=b:206046299
TEST=Repro case of b/206046299 does not cause any crash anymore.

Change-Id: I92b0b98f0d1c8092f0863c6ebd51bf5c417b2298
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3277960
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>