dns-proxy: Disable guests' redirection rules when device is not online
When device is not on an "online" state, DNS proxy redirection rules
should not be set for guests. Having the redirection rules breaks
DNS resolution for ARC.

An example of the case where DNS from ARC should be working when
device is not "online" can be observed when using ContentKeeper
application. When the device state is not yet "online", the resolver
for ARC proxy will no be running (as DoH would not be working). The
redirection rules in this state will make the DNS query fail.

BUG=b:206027483
TEST=dig google.com;
     sudo -u cups dig google.com;
     sudo -u chronos dig google.com;
TEST=Test DNS using ARC app;
TEST=Test DNS through Chrome;
TEST=tast run network.DNSProxy.*;
     tast run network.DNSProxyOverVPN.*;
TEST=Observe DNS redirection rules are not set up for guests when
     state is not "online";
TEST=Observe that /etc/resolv.conf does not point to the proxy when
     state is not "online";
TEST=Observe DNS redirection rules are not set up for guests when
     state is not "online" after re-login;
TEST=Observe that /etc/resolv.conf does not point to the proxy when
     state is not "online" after re-login;

Change-Id: Ib10239edd95b7771cb7a2618a23d219605700588
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3288410
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>