vm_tools: do not create /run/pvm with wrong ownership
Proper ownership of /run/pvm is pluginvm:pluginvm. Luckily the
dispatcher service unit was adjusting the ownership of the directory.

Let's move /run/pvm to the section of "other directories". It will still
be created with wrong ownership (root:root, as "-" means default
uid/gid in tmpfiles language), but it should be less confusing. The
dispatcher tmpfiles entry will need to make entry for this directory as
"force create" to make sure correct premissions/ownership is applied to
it).

BUG=None
TEST=tast run -buildbundle=crosint $DUT pita.Boot

Change-Id: I5279fd39155afe63a22a529126ad16a5cd9a8b10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3291868
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>