arc: Add mkdirat to forward-pstore arm policy
This was already added to the amd64 seccomp policy, but was missing
from the arm32 policy. With the new glibc 2.33 upgrade, we're seeing
seccomp failures in the field, leading to crashes in arcvm.

Crash link: http://shortn/_SVfX6dXtgB

This seccomp change attempts to fix the issue.

BUG=b:187795909
TEST=CQ

Change-Id: I46090ef304d9df30a5f011b7d4b940696a15706c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3290598
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Jordan R Abrahams <ajordanr@google.com>
Tested-by: Jordan R Abrahams <ajordanr@google.com>