biod: Get RecordId from file path when reading from disk
When record file exists but it's malformed then it might be not possible
to delete that file because currently RecordId is read from file.

With this patch, RecordId is obtained from path and later check is
performed to make sure that RecordId stored in the file is the same.
In case of mismatch RecordId from path is returned. That way, it is
possible to delete every invalid record.

BUG=b:184843581
TEST=FEATURES=test emerge-hatch biod

Change-Id: I05d58b129c1dd58e64ada21f8af2097e7750f6fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3291662
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Patryk Duda <patrykd@google.com>