shill: vpn: Wait for charon.vici connectable before running swanctl
Currently IPsecConnection will run swanctl right after it knows that the
vici socket file is created, but at that time the charon process may not
start listening on this socket, and thus the following execution of
swanctl may fail.

This patch adds the code to push off the execution of swanctl until the
vici socket is connectable. It is tested by calling connect() directly
on the socket. The mamimum number of attempts is limited, and it is very
likely that the new-added test code will not be triggered or be
triggered only once, so it will not affect the required time to
connected very much.

BUG=b:207298526
TEST=tast run $DUT network.VPNConnect.l2tp_ipsec_swanctl*; Checked that
  the new logic was triggered sometimes.

Change-Id: I5c09b551ba8004be4c578c192c9b17cacac87ab0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3295246
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>