attestation: Post OnOwnershipTakenSignal on the same thread
Currently, the OnOwnershipTakenSignal callback is called on the
tpm_manager thread, so we need to acquire lock before accessing is_ready
and local_data. But there are some places not locked correctly, and the
better way is to post the callback onto the same thread so that we don't
need the lock at all. Also, fix an incorrect return value in
TpmUtilityCommon::IsTpmReady() and modify relevant tests. Add a new test
to test the behavior that when TPM status is not owned or enabled,
IsReady should return false instead of true.

BUG=b:207321614
TEST=FEATURES=test emerge-$BOARD attestation

Change-Id: I69f22df66bc7657f01ca54bce2856c810fb77db2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3295247
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>