shill: passpoint: trigger interworking selection.
Shill now triggers interworking selection when a set of Passpoint
credentials is present and a passpoint compatible BSS is available
around. The logic implemented aims to reduce the number of selections,
thus avoid its cost when not required.

Each match between a set of credentials and a BSS triggers an event in
the WiFiDevice where the match is saved. At the end of the selection,
all the matches are processed and sent to the provider to (eventually)
populate services.

BUG=b:162106001
TEST=unit tests; create an AP and push the matching credentials to
Shill, then check the AP is matched correctly.

Change-Id: I1d87ac2aa3034d64bc1e824e281ec8399a97db2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3322995
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>