shill: cellular: defer connection if not yet registered
In cases where we have an attach APN, we may have some time
between the modem reaching the enabled state and actually becoming
registered. If we get a connect in that time period, we will fail
because it won't hit the scanning check and will instead hit the
not-registered error case. Set a pending connect so we come back
when registration is completed.

BUG=b:204847937,b:205882451
TEST=cellular tast/autotest suite

Change-Id: I8cae779f2f952d67c27235dbe88531549457fd56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3306718
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>