shill: metrics: remove shill's Daily/MonthlyChosen* metrics
This patch removes the implementation of shill UMA metrics
Network.Shill.{Daily,Monthly}Chosen{Fraction,Time}.*.

Inspection of the collected data shows that the metrics are not working
as was intended in the initial design and are not collecting any
information (see details in b/172213047).

Since the implementation is complex, not documented, and the original
authors have left the org, it is more sensible to remove these metrics,
redefine what was the original goal and see if it is still necessary to
measure.

BUG=b:172212984
BUG=b:172213047
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I14819e68c75ff81ca25c50e375967ce8b4db68ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3298647
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>