shill: passpoint: OI lists are now parsed a strings of decimal numbers
Chrome is not able to produce or consume uint64 values natively over
DBus because it is restricted to JSON native types, and uint32 is not
wide enough for accommodating all possible Passpoint OI (organization
identifier) values as per Passpoint specs and the supplicant
interworking implementation.

To work around this problem, OIs should be passed from Chrome to shill
encoded into strings in decimal representation.

Therefore this patch changes the processing of the OI lists passed to
shill over DBus (home OIs, required home OIs, roaming consortia OIs)
such that OIs are now passed as lists of decimal strings instead of
lists of uint64 values.

The internal memory representation of OIs as uint64 in the shill process
is not changed, and the storage format of OIs as decimal strings is also
not changed.

Changing the memory representation of OIs as uint64 could help remove
some format validation when retrieving Passoint credentials properties
from storage, but it would make the final conversion of OIs to the
supplicant format (hex strings of even length) more difficult.

BUG=b:162106001
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I254ecaae1c0e1a6257619a0a03b034d6bf8151bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3301135
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Damien Dejean <damiendejean@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>