hps: Fix i2c hal error reporting
Report the filename in quotes to make it more obvious if it is empty.
Explicitly fail on an empty filename. Report i2c rw errors to vlog,
because these are expected to happen while the device is being reset and
cause non error message log spam.

BUG=b:207722854
TEST=hps --bus=/dev/i2c-15 status
TEST=hps --bus /dev/i2c-15 status

Change-Id: I24b55d89207d961095b58bf7bbc7e61badd67b25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3301129
Commit-Queue: Evan Benn <evanbenn@chromium.org>
Tested-by: Evan Benn <evanbenn@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>