camera: fix camera3_module_test regression, part 2
Only check the number of cameras if it's not checked using the
cros_config info. On devices that have both USB and MIPI camera HALs,
one of the HAL can have zero built-in cameras and we should not fail the
test in this case if GetNumberOfCameras() reports 0.

BUG=b:207001080
TEST=Manually on device.

Change-Id: I16854d7be334e009bd8bed5e874f5b2a8a481abf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3301130
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>