rmad: Clear requests in test mode
Powerwash and cutoff requests are done by creating request files. In
test mode, the files are created in the test directory so they are not
picked up by the init script. This is intended to prevent actually
running powerwash or cutoff. Hence we additionally check these files in
test mode, and log that these requests are received but ignored.

BUG=None
TEST=FEATURES=test emerge-octopus rmad

Change-Id: I0bea2e3562578958fa2e3e56f9c63b3dc2285009
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3301894
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Gene Chang <genechang@google.com>