cryptohome: Prepare WebAuthn secret in CheckKey
In the design of WebAuthn, the WebAuthn secret needs to be unlocked when
the user successfully authenticates, including mounting and check key.
Currently, only mounting unlocks the secret. In order to support the
work of supporting WebAuthn in devices with TPM 1.2, the secret needs to
be unlocked when the user successfully performed CheckKey using PIN or
password.

After USS, the WebAuthn secret will be derived using another way, and
this change will not add any extra workload to USS. U2fd doesn't care
how the secret is derived too, as long as it gets the same value.

Add PrepareWebAuthnSecret functionality in userdataauth, and call it
when the user successfully performed a CheckKey using password or PIN.

BUG=b:184393647,b:205813697
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I5ed42b2ac05121f6a7c9aed00db2b45bfac659f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3301889
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>