libhwsec-foundation: Introduce hwsec's Error and StatusChain
Error is the object to use as a base for custom errors, as well as it
can be used directly. It provides a few customization end-points for
interaction with StatusChain.
StatusChain combines Status-like, unique_ptr-like, stack-like and
iterable-like properties with some additional behaviour to represents
a stack of polymorphic error objects.
StatusChain can be accessed as a status (ok()/error()), as a
uniuque_ptr(get()/reset()/->/*/nullptr conversion/comparison) or
iteratated upon (range()/const_range()). It also provides Wrap/Unwrap
methods to link another stack or remove the head element from the stack.

The internal structure of the object is a bit more complicated than it
could be in order to minimize the impact of changing the internal
representation of the stack on the structure of the rest of the class -
any required changes should be contained within Backend functions and
constructors, and not affect any other parts of the class.

BUG=b:207815464
TEST=FEATURES="test asan" USE="test asan" emerge-hatch libhwsec-foundation

Change-Id: I958fa41b7453b109ffbf71ab57f9b61b33f8a4bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3305118
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>