diagnostics: Switch reading UEFI SecureBoot in SystemFetcher to executor
The UEFI SecureBoot variable file on CloudReady 2.0 devices can only be
read by root, which cannot be done directly by healthd. It requires the
use of the cros_healthd executor which necessitated additions to the
executor implementation including the ability to request file contents
using the ReadFileToString method.

In addition, to enable access to the executor via IPC, SystemFetcher is
also refactored to be an asynchronous fetcher.

BUG=b:205774574
TEST=1. (cr) FEATURES=test USE="-cros-debug" emerge-reven diagnostics
     2. Deploy to DUT (a CloudReady 2.0 device)
     3. (DUT) cros-health-tool telem --category=system2
     4. (DUT) Verify there's no error in /var/log/messages

Change-Id: Ia5c315f747127ba3bb74754c76a0a0fa7949d8ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3305968
Reviewed-by: Kerker Yang <kerker@chromium.org>
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>
Tested-by: Denny Huang <dennyh@google.com>
Commit-Queue: Denny Huang <dennyh@google.com>