shill: passpoint: handle home/roaming consortium OIs
Translate the home and roaming consortium OIs from credentials to
supplicant properties. wpa_supplicant expects the OIs to be encoded as a
string of hexadecimal numbers. As the string is eventually converted to
an array of bytes, it must contain an even number of characters.

According to Passpoint specification, we should be able to support
multiple {home,required home} OIs, but it's not implemented in supplicant
yet. To provide the closest behavior, the following choice was made:
 - required home OI: OI list with more than one element are not
   supported. In such case the set of credentials is not pushed to
   supplicant. It avoids matching non compatible networks.
 - home OI: only the first element of the list is used. It may prevent
   matching some networks.

BUG=b:162106001
TEST=unit tests; connect to supplicant using wpa_cli, use list_creds and
get_cred to check "roaming_consortium" and "required_roaming_consortium".

Change-Id: I5c7504f1075a8be459392df162ebbbf381a87f77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3306406
Tested-by: Damien Dejean <damiendejean@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Damien Dejean <damiendejean@google.com>