fusebox: Add dbus::Response GetResponseErrno helper
The FuseBoxClient sends a DBUS message request to the FuseBoxServer,
to perform a file operation. The server's response to the request is
a dbus::Response* |response| that the client will process to extract
the file operation results. The client creates a dbus::MessageReader
initialized with |response|

  dbus::MessageReader reader(response);

and uses the MessageReader API to read values from the response. The
first value is a base::File::Error encoded as an int per CL:3306780.

Add a GetResponseErrno() utility to read that value and return it as
a system errno int. Add tests, including the nullptr |response| case
(the server did not respond to the request).

BUG=chromium:1249370
TEST=cros_workon_make --board=${BOARD} fusebox --test

Change-Id: Ia3cd2a71daa336c09c511ff8d801688090dbe2a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3307226
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>