power: DBus interface for toggling KBL off
Add separate DBus interface and state for toggling the keyboard
backlight on/off, which is what we do when the user presses the
keyboard key for toggling KBL.  Toggling off is functionally the same
as forcing off, except that the brightness change cause we send up is
USER_TOGGLED_ON/USER_TOGGLED_OFF.

The toggled-off flag is set to false on any user-initiated brightness
adjustment.

If the brightness change is a toggle, i.e. transition is INSTANT and
cause is TOGGLED_OFF/TOGGLED_ON, then we bypass the
percentage-has-changed check at the top of
KeyboardBacklightController::ApplyBrightnessPercent().

Migrate calls of Bind() to BindRepeating(), and Callback to
Repeatingcallback, to satisfy the presubmit checks.

Accompanying chromium-side change:

https://chromium-review.googlesource.com/c/chromium/src/+/3251651

BUG=b:194146863
TEST=Run it with the above Chromium CL

Change-Id: I33409684e88ab84bb7d2fe99f4e1dee1b5a4bd04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3309790
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Tested-by: Roger Tinkoff <rtinkoff@google.com>
Commit-Queue: Roger Tinkoff <rtinkoff@google.com>