camera: allow reading eeprom from i2c bus
The camera eeprom can be accessed from both the nvmem and i2c buses in
sysfs. For example, on Redrix the same camera eeprom file can be found
in the following two different paths:

- /sys/bus/i2c/devices/i2c-PRP0001:00/eeprom
- /sys/devices/pci0000:00/0000:00:15.2/i2c_designware.2/i2c-15/i2c-PRP0001:00/i2c-PRP0001:001/nvmem

Only the eeprom file in the i2c bus is set to group arc-camera. In order
for cros-camera to be able to read the module info from the eeprom, we
need to teach DeviceConfig to read from the i2c bus if the first read
attempt on the nvmem bus fails.

While we're at it, also add the PlatformCameraInfo::module_id() and
PlatformCameraInfo::sensor_id() getters to allow reusing the same id
strings in different places.

BUG=b:204958300
TEST=Verify that `cros-camera-tool modules list` and
`sudo -u arc-camera cros-camera-tool modules list` can both get the
module info.

Change-Id: I09f05c79164c52cc1a8112adc555dde9f97b77b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3312761
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>