sirenia: add manatee memory service
Start implementation of service that will run in the hypervisor and be
responsible for maintaining memory availability in the hypervisor. The
action of the service will be driven by a client in the CrOS guest. This
change adds the IPC plumbing, but doesn't handle any messages.

BUG=b:213962590
TEST=manual testing with crrev.com/c/3394436

Cq-Depend: chromium:3393112
Change-Id: Ic1bdb3e6496dbc7583c69e5996153af6d322b006
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3394433
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: David Stevens <stevensd@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>