login: Increase timeout for arc-boot-continue impulse
Currently we invoke the impulse synchronously with the default timeout
which is 25 seconds. We see reports in the production where arc-setup
is doing its work while session manager times out after 25 seconds
and kills the container.

In this change, we're changing the timeout from the default to 60
seconds. We will adjust the metric buckets in a separate change.

BUG=b:205032502
TEST=tast run DUT arc.Optin
TEST=cros_run_unit_tests --board=${BOARD} --packages chromeos-login

Change-Id: Ib22c0d6a90bff17910d428dd87e7977b30079f9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3313830
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>