vm_tools: cros_im: Manually define enums used in backend's API
This CL adds an explicit definition for zwp_text_input_v1_preedit_style,
which is used in IMContextBackend::Observer::SetPreedit(). Logically we
don't want the front end to be depending on the wayland text-input
headers as that should be abstracted by the backend.

Other enums defined in the text_input xml file will be similarly added
in the future as needed, for example for content hints.

BUG=chromium:1247564
TEST=ninja && GTK_IM_MODULE=cros gedit, input Japanese and see that
formatting (underlines) is as expected.

Change-Id: I9a8c10dbfa02ed01ea146f8a24f148628c5356cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3316170
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Tested-by: Timothy Loh <timloh@chromium.org>