vm_tools: cros_im: Add a libim_test_cros_gtk with mock Wayland backend
In preparation of actual tests, add a new build target which has the
Wayland headers mocked out when included directly. Note that the
frontend still has dependencies on actual Wayland headers transitively
via GTK (e.g. it needs to get the seat and surface to pass to the
backend), but this doesn't cause any issues.

As we have an entirely separate build target, we just define our own
implementations of the functions normally defined inline in
wayland-client.h and text-input-unstable-v1-client-protocol.h, using
the flag TEST_BACKEND to pick the appropriate header to use. The test
target does not depend on wayland-client, and should work on a regular
non-ChromeOS Linux machine (with a Wayland compositor).

The mock backend currently currently ignores Wayland requests, but a
subsequent CL will add functionality to define a test spec of expected
requests and events to fire in response.

BUG=chromium:1247564
TEST=ninja && GTK_IM_MODULE=cros gedit, Japanese input still works
TEST=ninja && GTK_IM_MODULE=test-cros gedit, no crashes (input doesn't
work)

Cq-Depend: chromium:3316170
Change-Id: Idb9b6147c9fba51eadfc79fa09bb590970a59f41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3317185
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Tested-by: Timothy Loh <timloh@chromium.org>