vm_tools: cros_im: Add basic tests (for commit_string)
Add a mechanism by which tests of the GTK IM module can be written, and
a couple of basic tests for commit_string. The test binary itself
behaves as a GTK application, so we can for example create a GTK text
field and have expectations of how the text buffer is edited or styled.
The mock backend is updated to support setting a list of Wayland
requests and events to expect and send, and receives environment
variable which indicates which expectations to use.

While it would be possible to have the main test code interact directly
with the mock backend (as the GTK IM module is just a shared library),
in the future we will want to add an XIM server, and having the backend
expectations configured separately would allow the same tests to be used
there as well.

There are a couple of caveats to running the tests, which will be
addressed in a later CL with a wrapper script: GTK_IM_MODULE must be set
appropriately, only one test can be run at a time (as GTK global state
can not be reset so we won't be able to re-initialize the IM module
with a different environment variable and set of expectations), and
expectation failures in the backend don't cause gtest to actually fail.

BUG=chromium:1247564
TEST=ninja; set up immodules.cache as per README.md, run ./cros_im_tests
for each test individually

Change-Id: I6ab49e6b218097eca36fff54a7ec88c08ba0a591
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3315025
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Tested-by: Timothy Loh <timloh@chromium.org>