sirenia: Switch sirenia_rpc over to &mut self.
In order to support trait nesting, &mut self is needed for all traits
and the client implementation is generalized over an invoker for a
procedure. This will allow the implementation to be reused in traits
that include other RPCs as super traits as long as the implement
Invoker<> for the proper type.

BUG=b:209442970
TEST=cargo test --workspace --all-features

Change-Id: I71b1c2d3aa255a84c0d7c1ac0ed87380da8d19e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3318729
Reviewed-by: Paramjit Oberoi <psoberoi@google.com>
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>