fusebox: add FUSE op GetAttr
Add FUSE op GetAttr: send a `stat` operation request to the server,
and handle its response to complete the FUSE GetAttr request.

"Cache" the struct stat in the Linux Kernel FUSE for (and following
SMBFS here) for 5 secs, kStatTimeoutSeconds [1].

[1] When the time expires, Kernel FUSE sends FUSE op Lookup to stat
file system entries, instead of GetAttr ops.

"Cache"? It's a cache of their time in the system: it's not a cache
of struct stats.

BUG=chromium:1244007
TEST=cros_workon_make --board=${BOARD} fusebox --test

Change-Id: I82fd4c4fe34fe42bb805b6eb5b438659ea6ee0cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3320127
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>