cryptohome: [USS] Fix missing continue in parsing
Fix a forgotten "continue" in one of the branches inside the
UserSecretStash parsing code, specifically the one that detects an
incorrect wrapped block algorithm.

This mistake was harmless, because during decryption we check the
algorithm again, but it's worth being fixed anyway. Also a test that
covers this piece of code was added.

BUG=b:195776293
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: If7f0cd981171992b1b83afa1aba1b2869fc05c49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3320227
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>