fusebox: add FUSE OpenDir op
Add FUSE op OpenDir: OpenDir is an optional FUSE op, used if a FUSE
user-space session wants to allocate a file handle for opendir(2).

This implementation will allocate file handles and does not need to
call the server for that. File handles are maintained in the client
and returned to Kernel FUSE, which returns them as file descriptors
(fd) to POSIX API callers.

How do we know that the given ino is a directory? Kernel FUSE sends
a GetAttr or Lookup op before an OpenDir op, to check the ino perms
and entry mode (stat) is a directory that can be opened.

BUG=chromium:1244007
TEST=cros_workon_make --board=${BOARD} fusebox --test

Change-Id: I9d0ca31073ed9941a206a8345761046bae45b108
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3322438
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>