cryptohome: Add GetHibernateSecret API
Following in the style of the WebAuthn secret, create a hibernate secret
derived from the FS encryption keys. This enables a mechanism where a
hibernate secret can be established based on user authentication data.
For now the secret itself is derived from the FS encryption keys.
Eventually this will be transformed into its own unique secret, with USS
as the backend storage for this secret.

BUG=b:204563517
TEST=emerge-volteer cryptohome

Change-Id: Ia03e5ca569b5c9f88f774a7d784e0ce6ad95af99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3335247
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>