fusebox: add FUSE ReleaseDir op
Add FUSE op ReleaseDir, for closedir(2). Undoes the work of FUSE op
OpenDir; close the file handle that the OpenDir op opened.

If the handle was not found, that is a possible symptom of a bug in
Kernel FUSE, or in our FUSE op implementations. Reply with EBADF in
that case to aid bug diagnosis.

BUG=chromium:1244007
TEST=cros_workon_make --board=${BOARD} fusebox --test

Change-Id: I238596f7bdc58f39a0e160f5f8dff975cefbe353
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3325801
Commit-Queue: Noel Gordon <noel@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>