shill: vpn: Add DeviceInfo::CreateXFRMInterface()
This patch adds a CreateXFRMInterface() method in the DeviceInfo class,
which will be used by the IPsecConnection class to create the XFRM
interface for IKEv2 connection. Creating an XFRM interface needs to
passing IFLA_INFO_DATA property in the RTNL message, and thus this patch
also changes the interface of RTNLHandler.

BUG=b:210064468
TEST=unit_tests

Change-Id: Ia1a47f5b3e78eccf05e4f3d74af05ed6a54f8754
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3334319
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>