patchpanel: Stop multicast forwarding on shill device removed
When an interface is removed, multicast forwarder should stop
the forwarder of the interface. Without stopping, the forwarder
will not be able to bind to the correct interface when the device
is reconnected, turning the forwarder into a defunct state. This
patch stops the forwarder when a shill device is removed.

While at it, this patch also change the forwarding interfaces cache
for ndproxy to mirror the new multicast forwarder's cache.

IPv6 forwarding is also stopped on shill device removed.

BUG=b:209910496
BUG=b:166895274
TEST=Unplug ethernet;
     Replug ethernet;
     tast run <DUT> arc.MulticastForwarder; # passes

Change-Id: Ie0155028db69bfb4f56b9a96f98740175ad0a363
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3331989
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>