shill: Emit current MAC address to be displayed
This change is required to properly display a randomized MAC address.
Whenever connection attempt asks to update MAC address, we also
trigger emission of the address, so it gets updated in Chrome.

Works in Chrome out-of-the-box, does not require changes there.

BUG=b:181234131
TEST=Enable MAC Address Randomization for connection experimental flag;
connect to a new network, observe updated MAC Address in Chrome dialog.

Change-Id: Idc20a7282783d8b2483362cdb5309b65129fe388
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3330471
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Jacek Siuda <jsiuda@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Jacek Siuda <jsiuda@google.com>