ml: Retroactively fixed test_suggester.mojom versioning
**This CL will need to be chumped** in order to bypass the stable mojom
presubmit.

This mojom had a non-backards-compatible change made to a [Stable]
interface before a presubmit check was added to prevent that.  This CL
retroactively adds the appropriate versioning info to that change such
that it would have passed the presubmit. This also happens to bring the
mojom back in sync with the chromium side, where the versioning info was
already added (presubably because the presubmit already existed on that
side and required it).

BUG=chromium:1283517
TEST=Run roll_mojoms.sh, no changes are proposed.

Change-Id: Ied9058271f3a6d4fbff6dfe2c6a50c8aae57b5b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3337019
Tested-by: Andrew Moylan <amoylan@chromium.org>
Reviewed-by: Xinglong Luan <alanlxl@chromium.org>
Tested-by: Curtis McMullan <curtismcmullan@chromium.org>